+++ bpm-engine-api/src/main/java/io/takari/bpm/model/SubProcess.java
@@ -1,5 +1,6 @@
 package io.takari.bpm.model;
 
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Set;
@@ -11,6 +12,10 @@ public class SubProcess extends ProcessDefinition {
     private final boolean useSeparateContext;
     private final Set<VariableMapping> outVariables;
 
+    public SubProcess(String id, AbstractElement... children) {
+        this(id, Arrays.asList(children));
+    }
+
     public SubProcess(String id, Collection<AbstractElement> children) {
         this(id, false, Collections.emptySet(), children);
     }
@@ -18,11 +23,11 @@ public class SubProcess extends ProcessDefinition {
     public SubProcess(String id, boolean useSeparateContext, Collection<AbstractElement> children) {
         this(id, useSeparateContext, null, children);
     }
-    
+
     /**
      * @param id
      * @param useSeparateContext if {@code false} use the context of a parent process. All changes to variables will be
-     *                           visible to sibling and following processes and tasks. If {@code true} all changes, except 
+     *                           visible to sibling and following processes and tasks. If {@code true} all changes, except
      *                           those specified in outs, will be discarded.
      * @param outVariables if {@code useSeparateContext} is set to {@code true}, use these variable mappings as out parameters,
      *                     ignored if {@code useSeparateContext} is {@code false}
@@ -37,7 +42,7 @@ public class SubProcess extends ProcessDefinition {
     public boolean isUseSeparateContext() {
         return useSeparateContext;
     }
-    
+
     public Set<VariableMapping> getOutVariables() {
         return outVariables;
     }