+++ bpm-engine-impl/src/main/java/io/takari/bpm/Configuration.java
@@ -5,8 +5,8 @@ import java.io.Serializable;
 public class Configuration implements Serializable {
 
     private boolean throwExceptionOnErrorEnd = false;
-    private boolean throwExceptionOnUnhandledBpmnError = true;
     private boolean avoidDefinitionReloadingOnCall = true;
+    private UnhandledBpmnErrorStrategy unhandledBpmnErrorStrategy = UnhandledBpmnErrorStrategy.PROPAGATE;
 
     public boolean isThrowExceptionOnErrorEnd() {
         return throwExceptionOnErrorEnd;
@@ -21,16 +21,31 @@ public class Configuration implements Serializable {
     }
 
     public boolean isThrowExceptionOnUnhandledBpmnError() {
-        return throwExceptionOnUnhandledBpmnError;
+        return unhandledBpmnErrorStrategy == UnhandledBpmnErrorStrategy.EXCEPTION;
     }
 
     /**
      * Throw an exception on an unhandled {@BpmnError} (e.g. when error occurs in a subprocess without a boundary error
      * event).
      * @param throwExceptionOnUnhandledBpmnError
+     * @deprecated use {@link #setUnhandledBpmnErrorStrategy(UnhandledBpmnErrorStrategy)}
      */
+    @Deprecated
     public void setThrowExceptionOnUnhandledBpmnError(boolean throwExceptionOnUnhandledBpmnError) {
-        this.throwExceptionOnUnhandledBpmnError = throwExceptionOnUnhandledBpmnError;
+        this.unhandledBpmnErrorStrategy = UnhandledBpmnErrorStrategy.EXCEPTION;
+    }
+
+    public UnhandledBpmnErrorStrategy getUnhandledBpmnErrorStrategy() {
+        return unhandledBpmnErrorStrategy;
+    }
+
+    /**
+     * Determines how to process unhandled BPMN errors (e.g. errors in subprocesses without boundary error events).
+     * By default, unhandled errors are propagated to a higher level.
+     * @param unhandledBpmnErrorStrategy
+     */
+    public void setUnhandledBpmnErrorStrategy(UnhandledBpmnErrorStrategy unhandledBpmnErrorStrategy) {
+        this.unhandledBpmnErrorStrategy = unhandledBpmnErrorStrategy;
     }
 
     public boolean isAvoidDefinitionReloadingOnCall() {
@@ -44,5 +59,24 @@ public class Configuration implements Serializable {
      */
     public void setAvoidDefinitionReloadingOnCall(boolean avoidDefinitionReloadingOnCall) {
         this.avoidDefinitionReloadingOnCall = avoidDefinitionReloadingOnCall;
+    }
+
+    public enum UnhandledBpmnErrorStrategy {
+
+        /**
+         * Throw an {@code {@link io.takari.bpm.api.ExecutionException}} immediately.
+         */
+        EXCEPTION,
+
+        /**
+         * Propagate an error to a higher level of a process' stack (similar to Java's exception propagation).
+         */
+        PROPAGATE,
+
+        /**
+         * Exit a subprocess, ignore an error and continue the execution.
+         * While this doesn't make much sense, it left for the compatibility with old versions of the engine.
+         */
+        IGNORE
     }
 }